+++ core/src/Content/Soldat/SoldatFlame.java
@@ -3,7 +3,7 @@ import com.mygdx.game.main.Main;
 import com.mygdx.game.soldat.Soldat;
 
 public class SoldatFlame extends Soldat {
-    public SoldatFlame(double x, double y){
+    public SoldatFlame(float x, float y){
         this.name = "bull";
         this.x = x;
         this.y = y;
