+++ core/src/Content/Soldat/SoldatPacket.java
@@ -1,7 +1,7 @@
 package Content.Soldat;
 
 public class SoldatPacket {
-    public double x,y,rotation;
+    public float x,y,rotation;
     public String name;
     public byte team;
     //public byte clear_sost;
